+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.1] - 2021-04-05
+### Fixed
+- Add the user multiple time in to the same VO(COU)
+
 ## [0.3.0] - 2021-04-02
 ### Added
 - Made Certificate Model Name, Certificate Subject DN column name, Certificate Issuer DN column name configurable